+++ src/main/java/raft/server/RaftStatusSnapshot.java
@@ -10,9 +10,9 @@ import java.util.List;
 public class RaftStatusSnapshot {
     public static final RaftStatusSnapshot emptyStatus = new RaftStatusSnapshot();
 
-    private int term;
-    private int commitIndex;
-    private int appliedIndex;
+    private long term;
+    private long commitIndex;
+    private long appliedIndex;
     private String leaderId;
     private State state;
     private List<String> peerNodeIds;
@@ -26,11 +26,11 @@ public class RaftStatusSnapshot {
         this.peerNodeIds = Collections.emptyList();
     }
 
-    public void setTerm(int term) {
+    public void setTerm(long term) {
         this.term = term;
     }
 
-    public void setCommitIndex(int commitIndex) {
+    public void setCommitIndex(long commitIndex) {
         this.commitIndex = commitIndex;
     }
 
@@ -46,11 +46,11 @@ public class RaftStatusSnapshot {
         this.peerNodeIds = peerNodeIds;
     }
 
-    public int getTerm() {
+    public long getTerm() {
         return term;
     }
 
-    public int getCommitIndex() {
+    public long getCommitIndex() {
         return commitIndex;
     }
 
@@ -62,11 +62,11 @@ public class RaftStatusSnapshot {
         return state;
     }
 
-    public int getAppliedIndex() {
+    public long getAppliedIndex() {
         return appliedIndex;
     }
 
-    public void setAppliedIndex(int appliedIndex) {
+    public void setAppliedIndex(long appliedIndex) {
         this.appliedIndex = appliedIndex;
     }
